pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "learning-organoid-programming"
version = "0.1.0"
description = "Study notes and experiments with cl-sdk for organoid computing"
license = { file = "LICENSE.txt" }
requires-python = ">=3.12"
dependencies = [
    "cl-sdk",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.0.0",
    "notebook>=7.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "nbqa",
    "nbstripout",
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-versions = ["py312"]
